// types/listingType.ts
export interface ListingType{
    id: string;
    name: string;
    image: string;
    description: string;
    rating: number;
    price: string;
    duration: string;
    location: string;
    category: string;
    previewImage1: string;
    previewImage2: string;
    previewImage3: string;
    previewImage4: string;

}